Moved authenticator-options file to more suitable location and added linters
