+++ backend/src/apps/users/domain/services/authenticator/index.js
@@ -1,5 +1,5 @@
 import buildAuthenticator from "./authenticator.js";
-import authenticatorOptions from "../../../authentication/authenticator-options.js";
+import authenticatorOptions from "./authenticator-options.js";
 import configuredPassportInstance from "../../../authentication/passport-setup/index.js";
 
 const authenticator = buildAuthenticator({
