+++ package.json
@@ -11,9 +11,11 @@
   "dependencies": {
     "@supabase/supabase-js": "^2.21.0",
     "axios": "^1.3.6",
+    "eslint": "^8.44.0",
     "fs": "^0.0.1-security",
     "pdf-img-convert": "^1.2.1",
     "pdf2pic": "^2.1.4",
+    "prettier": "^2.8.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.10.0",
@@ -23,6 +25,7 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react": "^3.1.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "tailwindcss": "^3.3.1",
     "vite": "^4.2.0"
   }
